# daemon/api/api.yaml
openapi: 3.0.0
info:
  title: opensvc agent api
  version: 4.2.0
paths:
...
  /api/node/name/{nodename}/instance/path/{namespace}/{kind}/{name}/resource/file:
    get:
      operationId: GetInstanceResourceFile
      description: |
        For internal use only.

        Returns the content of the requested file as the response body, with the file’s metadata (statinfo) included in the response headers. This operation is only performed if the specified file path is verified as exposed by an instance resource.
      parameters:
        - $ref: '#/components/parameters/inPathNodeName'
        - $ref: '#/components/parameters/inPathNamespace'
        - $ref: '#/components/parameters/inPathKind'
        - $ref: '#/components/parameters/inPathName'
        - $ref: '#/components/parameters/inQueryResourceFileName'
        - $ref: '#/components/parameters/inQueryResourceFileRid'
      security:
        - basicAuth: []
        - bearerAuth: []
      responses:
        200:
          description: OK
          headers:
            OM-Group:
              type: string
            OM-Last-Modified:
              type: string
              format: date-time
              pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(\.\d{1,9})?Z?$'
            OM-Perm:
              type: string
            OM-User:
              type: string
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        400:
          $ref: '#/components/responses/400'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        500:
          $ref: '#/components/responses/500'
      tags:
        - node / instance / svc
        - node / instance / vol
      x-internal: true
...
components:
  schemas:
...
    Kind:
      x-go-type: naming.Kind
      x-go-type-import:
        path: github.com/opensvc/om3/v3/core/naming
      type: string
      enum:
        - svc
        - vol
        - cfg
        - sec
        - usr
        - nscfg
        - ccfg
...
    Problem:
      type: object
      properties:
        detail:
          description: |
            A human-readable explanation specific to this occurrence of the
            problem.
          type: string
        status:
          type: integer
          description: |
            The HTTP status code ([RFC7231], Section 6) generated by the
            origin server for this occurrence of the problem.
        title:
          type: string
          description: |
            A short, human-readable summary of the problem type.  It SHOULD
            NOT change from occurrence to occurrence of the problem, except
            for purposes of localization (e.g., using proactive content
            negotiation; see [RFC7231], Section 3.4).
      required:
        - detail
        - status
        - title
...
  parameters:
...
    inPathKind:
      in: path
      name: kind
      required: true
      schema:
        $ref: '#/components/schemas/Kind'
...
    inPathName:
      in: path
      name: name
      required: true
      schema:
        type: string
    inPathNamespace:
      in: path
      name: namespace
      required: true
      schema:
        type: string
        example: test
    inPathNodeName:
      in: path
      name: nodename
      required: true
      schema:
        type: string
        description: |
          The cluster node to relay the request to. `_` and `localhost` refer tp
          the node that received the request.
        example: localhost
...
    inQueryResourceFileName:
      in: query
      name: name
      description: The path of a file explicitely shared by a resource driver that the requester wants to receive the data of.
      required: true
      schema:
        type: string
    inQueryResourceFileRid:
      in: query
      name: rid
      description: The rid that is expected to expose the requested file path.
      required: true
      schema:
        type: string
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic

    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    '400':
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
    '401':
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
    '403':
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    '500':
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
...